Finish renaming of counter -> count
