+++ apps/web/src/lib/schemas/countLoginSchema.ts
@@ -2,6 +2,6 @@ import { z } from 'zod';
 
 import { pubKeySchema } from './pubKeyIdPairSchema';
 
-export const counterLoginSchema = z.object({
+export const countLoginSchema = z.object({
 	pubKey: pubKeySchema
 });
